--- src/cert_utils.rs
@@ -1,329 +1,0 @@
-#![cfg(all(feature = "scard", target_os = "windows"))]
-#![expect(clippy::missing_safety_doc)]
-
-use std::ffi::c_void;
-use std::ptr::{null, null_mut};
-use std::slice::from_raw_parts;
-
-use picky_asn1_x509::Certificate;
-use sha1::{Digest, Sha1};
-use windows_sys::Win32::Security::Cryptography::{
-    CertCloseStore, CertEnumCertificatesInStore, CertFreeCertificateContext, CertOpenStore, CryptAcquireContextW,
-    CryptDestroyKey, CryptGetKeyParam, CryptGetProvParam, CryptGetUserKey, CryptReleaseContext, AT_KEYEXCHANGE,
-    CERT_STORE_PROV_SYSTEM_W, CERT_SYSTEM_STORE_CURRENT_USER_ID, CERT_SYSTEM_STORE_LOCATION_SHIFT, CRYPT_FIRST,
-    CRYPT_NEXT, CRYPT_SILENT, KP_CERTIFICATE, PP_ENUMCONTAINERS, PP_SMARTCARD_READER, PROV_RSA_FULL,
-};
-
-// UTF-16 encoded "Microsoft Base Smart Card Crypto Provider\0"
-const CSP_NAME_W: &[u8] = &[
-    77, 0, 105, 0, 99, 0, 114, 0, 111, 0, 115, 0, 111, 0, 102, 0, 116, 0, 32, 0, 66, 0, 97, 0, 115, 0, 101, 0, 32, 0,
-    83, 0, 109, 0, 97, 0, 114, 0, 116, 0, 32, 0, 67, 0, 97, 0, 114, 0, 100, 0, 32, 0, 67, 0, 114, 0, 121, 0, 112, 0,
-    116, 0, 111, 0, 32, 0, 80, 0, 114, 0, 111, 0, 118, 0, 105, 0, 100, 0, 101, 0, 114, 0, 0, 0,
-];
-const CSP_NAME: &str = "Microsoft Base Smart Card Crypto Provider";
-
-// https://learn.microsoft.com/en-us/windows/win32/seccrypto/hcryptprov
-pub type HCRYPTPROV = usize; // ULONG_PTR
-                             // https://learn.microsoft.com/en-us/windows/win32/seccrypto/hcryptkey
-pub type HCRYPTKEY = usize; // ULONG_PTR
-
-use crate::{Error, ErrorKind, Result};
-
-#[instrument(level = "trace", ret)]
-unsafe fn find_raw_cert_by_thumbprint(thumbprint: &[u8], cert_store: *mut c_void) -> Result<Vec<u8>> {
-    let mut certificate = CertEnumCertificatesInStore(cert_store, null_mut());
-
-    while !certificate.is_null() {
-        let cert_der = from_raw_parts((*certificate).pbCertEncoded, (*certificate).cbCertEncoded as usize);
-        let mut sha1 = Sha1::new();
-        sha1.update(cert_der);
-        let cert_thumbprint = sha1.finalize().to_vec();
-
-        if cert_thumbprint == thumbprint {
-            CertFreeCertificateContext(certificate);
-
-            return Ok(cert_der.to_vec());
-        }
-
-        let next_certificate = CertEnumCertificatesInStore(cert_store, certificate);
-
-        certificate = next_certificate;
-    }
-
-    Err(Error::new(
-        ErrorKind::InternalError,
-        "Cannot find appropriate device certificate",
-    ))
-}
-
-fn open_user_cert_store() -> Result<*mut c_void> {
-    // "My\0" encoded as a wide string.
-    // More info: https://docs.microsoft.com/en-us/windows/win32/api/wincrypt/nf-wincrypt-certopenstore#remarks
-    let my: [u16; 3] = [77, 121, 0];
-    // SAFETY:
-    // * constant parameters are taken from the `windows_sys` crate. Thus, they are valid;
-    // * `dwEncodingType` and `hCryptProv` are allowed to be zero by documentation;
-    // * `my` is as valid wide C string.
-    let cert_store = unsafe {
-        CertOpenStore(
-            CERT_STORE_PROV_SYSTEM_W,
-            0,
-            0,
-            CERT_SYSTEM_STORE_CURRENT_USER_ID << CERT_SYSTEM_STORE_LOCATION_SHIFT,
-            my.as_ptr() as *const _,
-        )
-    };
-
-    if cert_store.is_null() {
-        return Err(Error::new(
-            ErrorKind::InternalError,
-            "Cannot initialize certificate store: permission denied",
-        ));
-    }
-
-    Ok(cert_store)
-}
-
-#[instrument(level = "trace", ret)]
-pub fn extract_raw_certificate_by_thumbprint(thumbprint: &[u8]) -> Result<Vec<u8>> {
-    let cert_store = open_user_cert_store()?;
-    // SAFETY: `open_user_cert_store` returns valid store handle.
-    let cert = unsafe { find_raw_cert_by_thumbprint(thumbprint, cert_store) }.inspect_err(|_err| {
-        // SAFETY: `open_user_cert_store` returns valid store handle that needs to be closed.
-        if unsafe { CertCloseStore(cert_store, 0) } == 0 {
-            warn!("could not close the certificate store");
-        }
-    })?;
-
-    // SAFETY: `open_user_cert_store` returns valid store handle that needs to be closed.
-    if unsafe { CertCloseStore(cert_store, 0) } == 0 {
-        return Err(Error::new(
-            ErrorKind::InternalError,
-            "could not close the certificate store",
-        ));
-    }
-
-    Ok(cert)
-}
-
-#[instrument(level = "trace", ret)]
-pub fn extract_certificate_by_thumbprint(thumbprint: &[u8]) -> Result<(Vec<u8>, Certificate)> {
-    let raw_cert = extract_raw_certificate_by_thumbprint(thumbprint)?;
-
-    Ok((raw_cert.to_vec(), picky_asn1_der::from_bytes(&raw_cert)?))
-}
-
-#[instrument(level = "trace", ret)]
-unsafe fn acquire_key_container_context(key_container_name: &str) -> Result<HCRYPTPROV> {
-    let container_name = key_container_name
-        .encode_utf16()
-        .chain(std::iter::once(0))
-        .flat_map(|v| v.to_le_bytes())
-        .collect::<Vec<_>>();
-    let mut crypt_context_handle = HCRYPTPROV::default();
-
-    if CryptAcquireContextW(
-        &mut crypt_context_handle,
-        container_name.as_ptr() as *const _,
-        CSP_NAME_W.as_ptr() as *const _,
-        PROV_RSA_FULL,
-        CRYPT_SILENT,
-    ) == 0
-    {
-        return Err(Error::new(
-            ErrorKind::InternalError,
-            "Cannot acquire crypt context handle.",
-        ));
-    }
-
-    Ok(crypt_context_handle)
-}
-
-#[instrument(level = "trace", ret)]
-unsafe fn get_reader_name(crypt_context_handle: HCRYPTPROV) -> Result<String> {
-    let mut reader_buff_len = 0;
-    if CryptGetProvParam(
-        crypt_context_handle,
-        PP_SMARTCARD_READER,
-        null_mut(),
-        &mut reader_buff_len,
-        0,
-    ) == 0
-    {
-        return Err(Error::new(ErrorKind::InternalError, "Cannot get reader name."));
-    }
-
-    let mut reader_buff = vec![0; reader_buff_len as usize];
-    if CryptGetProvParam(
-        crypt_context_handle,
-        PP_SMARTCARD_READER,
-        reader_buff.as_mut_ptr(),
-        &mut reader_buff_len,
-        0,
-    ) == 0
-    {
-        return Err(Error::new(ErrorKind::InternalError, "Cannot get reader name."));
-    }
-
-    // remove null byte
-    reader_buff.pop();
-
-    String::from_utf8(reader_buff)
-        .map_err(|_| Error::new(ErrorKind::InternalError, "reader name is not valid UTF-8 text"))
-}
-
-#[instrument(level = "trace", ret)]
-pub unsafe fn get_key_container_certificate(crypt_context_handle: HCRYPTPROV) -> Result<Certificate> {
-    let mut key = HCRYPTKEY::default();
-
-    if CryptGetUserKey(crypt_context_handle, AT_KEYEXCHANGE, &mut key) == 0 {
-        return Err(Error::new(ErrorKind::InternalError, "Cannot acquire key handle."));
-    }
-
-    let mut cert_data_len = 0;
-    if CryptGetKeyParam(key, KP_CERTIFICATE, null_mut(), &mut cert_data_len, 0) == 0 {
-        CryptDestroyKey(key);
-        return Err(Error::new(ErrorKind::InternalError, "Cannot get certificate data len."));
-    }
-
-    let mut cert_data = vec![0; cert_data_len as usize];
-    if CryptGetKeyParam(key, KP_CERTIFICATE, cert_data.as_mut_ptr(), &mut cert_data_len, 0) == 0 {
-        CryptDestroyKey(key);
-        return Err(Error::new(ErrorKind::InternalError, "Cannot get certificate data."));
-    }
-
-    CryptDestroyKey(key);
-
-    Ok(picky_asn1_der::from_bytes(&cert_data)?)
-}
-
-#[derive(Debug)]
-pub struct SmartCardInfo {
-    pub key_container_name: String,
-    pub reader_name: String,
-    pub csp_name: String,
-    pub certificate: Certificate,
-}
-
-// This function gathers the smart card information like reader name, key container name,
-// and so on using the provided certificate serial number.
-// It iterates over existing key containers and tries to find a suitable reader name and key container.
-// The similar approach is implemented in the FreeRDP for the smart card information gathering:
-// https://github.com/FreeRDP/FreeRDP/blob/56324906a2d5b2538675e2f10b9f1ffe4a27de79/libfreerdp/core/smartcardlogon.c#L616
-#[instrument(level = "trace", ret)]
-pub fn finalize_smart_card_info(cert_serial_number: &[u8]) -> Result<SmartCardInfo> {
-    let mut crypt_context_handle = HCRYPTPROV::default();
-    // SAFETY:
-    // * `phProv` is constructed using Rust references. This it is valid.
-    // * `szContainer` can be null according to the documentation.
-    // * all other parameters are hardcoded constants that are valid.
-    if unsafe {
-        CryptAcquireContextW(
-            &mut crypt_context_handle,
-            null(),
-            CSP_NAME_W.as_ptr() as *const _,
-            PROV_RSA_FULL,
-            CRYPT_SILENT,
-        )
-    } == 0
-    {
-        return Err(Error::new(
-            ErrorKind::InternalError,
-            "Cannot acquire crypt context handle.",
-        ));
-    }
-
-    let mut key_container_name_len = 0;
-    let mut is_first = true;
-    loop {
-        // SAFETY:
-        // * `crypt_context_handle` is obtained from the successful `CryptAcquireContextW` function call.
-        // * `pbData` can be null according to the documentation.
-        // * `key_container_name_len` is constructed using Rust references. This it is valid.
-        // * all other parameters are hardcoded constants that are valid.
-        if unsafe {
-            CryptGetProvParam(
-                crypt_context_handle,
-                PP_ENUMCONTAINERS,
-                null_mut(),
-                &mut key_container_name_len,
-                if is_first { CRYPT_FIRST } else { CRYPT_NEXT },
-            )
-        } == 0
-        {
-            break;
-        }
-
-        let mut key_container_name = vec![0; key_container_name_len as usize];
-
-        // SAFETY:
-        // * `crypt_context_handle` is obtained from the successful `CryptAcquireContextW` function call.
-        // * `key_container_name` is a Rust vector buffer of valid length. Its len was previously obtained using `CryptGetProvParam`.
-        // * `key_container_name_len` is constructed using Rust references. This it is valid.
-        // * all other parameters are hardcoded constants that are valid.
-        if unsafe {
-            CryptGetProvParam(
-                crypt_context_handle,
-                PP_ENUMCONTAINERS,
-                key_container_name.as_mut_ptr(),
-                &mut key_container_name_len,
-                if is_first { CRYPT_FIRST } else { CRYPT_NEXT },
-            )
-        } == 0
-        {
-            break;
-        }
-        let mut key_container_name = String::from_utf8(key_container_name).unwrap();
-        // remove null char
-        key_container_name.pop();
-
-        let context = if let Ok(context) = unsafe { acquire_key_container_context(&key_container_name) } {
-            context
-        } else {
-            continue;
-        };
-
-        if let Ok(certificate) = unsafe { get_key_container_certificate(context) } {
-            if certificate.tbs_certificate.serial_number.0 == cert_serial_number {
-                let reader_name = unsafe { get_reader_name(crypt_context_handle) };
-
-                // SAFETY: the `crypt_context_handle` was obtained using successful `CryptAcquireContextW` function call.
-                // `dwFlags` parameter is reserved for future use and must be zero.
-                if unsafe { CryptReleaseContext(crypt_context_handle, 0) } == 0 {
-                    return Err(Error::new(
-                        ErrorKind::InternalError,
-                        "could not release the crypto context",
-                    ));
-                }
-
-                let reader_name = match reader_name {
-                    Ok(reader_name) => reader_name,
-                    Err(err) => {
-                        error!(?err);
-                        continue;
-                    }
-                };
-
-                return Ok(SmartCardInfo {
-                    key_container_name,
-                    reader_name,
-                    certificate,
-                    csp_name: CSP_NAME.to_owned(),
-                });
-            }
-        }
-
-        is_first = false;
-    }
-
-    // SAFETY: the `crypt_context_handle` was obtained using successful `CryptAcquireContextW` function call.
-    // `dwFlags` parameter is reserved for future use and must be zero.
-    if unsafe { CryptReleaseContext(crypt_context_handle, 0) } == 0 {
-        return Err(Error::new(
-            ErrorKind::InternalError,
-            "could not release the crypto context",
-        ));
-    }
-
-    Err(Error::new(ErrorKind::InternalError, "Cannot get smart card info"))
-}
